backend/requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Firebase Admin SDK
firebase-admin>=6.3.0

# Encryption
pycryptodome>=3.19.0

# QR Code
qrcode>=7.4.2
Pillow>=10.1.0

# Environment variables
python-dotenv>=1.0.0

# Production server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Utilities
python-dateutil>=2.8.2
